iceberg_to_ch/requirements.txt
pyspark==3.4.0
clickhouse-driver==0.2.0
python-dotenv==1.0.0
